ms-alpha/requirements.txt
# App
Flask==2.2.3
Werkzeug==2.2.2

# Database
python-dotenv==0.20.0
mysql-connector-python==8.0.29
sqlalchemy==1.4.47
Flask-SQLAlchemy==2.5.1

# CORS
flask-cors==3.0.10

# Circuit breaker
pybreaker==0.7.0